reorganise existing productPrice tests
